% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCodelist.R
\name{buildCodelist}
\alias{buildCodelist}
\title{Add RDF data cube codelist to RDF store}
\usage{
buildCodelist(store, prefixlist, obsData, codeType, nciDomainValue, dimName,
  underlDataSetName = NULL, remote.endpoint = NULL,
  extension.rrdfqbcrnd0 = FALSE)
}
\arguments{
\item{obsData}{Data Frame with the data for which the code list is to be generated}

\item{codeType}{Character "DATA" or "SDTM".
"DATA" to derive code list from the data.
"SDTM" to derive the code list from the rdf.cdisc.org documentation using a SPARQL query}

\item{nciDomainValue}{When codetype="SDTM" the nciDomain used for identifying the codelist}

\item{dimName}{the name of the dimension - for codeType="DATA" the name of the variable in the data frame ObsData}

\item{underlDataSetName}{underlying data set name. Used for finding name for D2RQ propertybridge. If NULL then not used.}

\item{remote.endpoint}{Used when codetype="SDTM" to give the URL for the remote endpoint. If NULL then the local rdf.cdisc.store from the environment is used.}

\item{extension.rrdfqbcrnd0}{If TRUE then rrdfqbcrnd0 specific values will be added to the generated cube}
}
\value{
Alway TRUE - to be corrected
}
\description{
This function could be split into two function corresponding to the usage.
}
\author{
Tim Williams, Marc Andersen
}

