% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCodelist.SDTM.R
\name{buildCodelist.SDTM}
\alias{buildCodelist.SDTM}
\title{Add codelist derived from SDTM terminology to RDF data cube SDTM derived codelist to RDF store.
The code list from the rdf.cdisc using a SPARQL query.}
\usage{
buildCodelist.SDTM(store, prefixlist, obsData, nciDomainValue, dimName,
  underlDataSetName = NULL, remote.endpoint = NULL)
}
\arguments{
\item{store}{The name for the rrdf store}

\item{prefixlist}{R list of the prefixes}

\item{obsData}{Data Frame with the data for which the code list is to be generated}

\item{nciDomainValue}{the nciDomain used for identifying the codelist}

\item{dimName}{the name of the dimension}

\item{underlDataSetName}{underlying data set name. Used for finding name for D2RQ propertybridge. If NULL then not used.}

\item{remote.endpoint}{Provides the URL for the remote endpoint. If NULL then the local rdf.cdisc.store from the environment is used.}
}
\value{
Alway TRUE - to be corrected
}
\description{
Add codelist derived from SDTM terminology to RDF data cube SDTM derived codelist to RDF store.
The code list from the rdf.cdisc using a SPARQL query.
}
\author{
Tim Williams, Marc Andersen
}

