% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.prefixlist.from.df.R
\name{Get.prefixlist.from.df}
\alias{Get.prefixlist.from.df}
\title{Return list of prefixes}
\usage{
Get.prefixlist.from.df(prefixes)
}
\arguments{
\item{prefixes}{A data.frame with column prefix and namespace}
}
\value{
The list with member names prefixUPPERCASEPREFIX and namespace as value
TODO(mja): this should be changed - at least the naming, or use another data structure
}
\description{
Return list of prefixes
}

