% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.rq.prefix.df.R
\name{Get.rq.prefix.df}
\alias{Get.rq.prefix.df}
\title{Create prefix definition for SPARQL query from}
\usage{
Get.rq.prefix.df(prefix.df)
}
\arguments{
\item{prefix.df}{prefixes as data.frame with column prefix and namespace}
}
\value{
String containing prefix statements for SPARQL query
}
\description{
The function defines the domain specifif prefixes used for the rdf
data cube for the Clinical Research and Development Analysis
Results Data
}

