% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAttributesSparqlQuery.R
\name{GetAttributesSparqlQuery}
\alias{GetAttributesSparqlQuery}
\title{SPARQL query for attributes in RDF data cube}
\usage{
GetAttributesSparqlQuery(forsparqlprefix)
}
\arguments{
\item{forsparqlprefix}{string with prefix defintions}
}
\value{
SPARQL query
}
\description{
SPARQL query for attributes in RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetCodeListSparqlQueryFull}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetCodeListSpecSparqlQuery}},
  \code{\link{GetComponentSparqlQuery}},
  \code{\link{GetDimAttrSparqlQuery}},
  \code{\link{GetDimensionsSparqlQuery}},
  \code{\link{GetDimsubsetDescSparqlQuery}},
  \code{\link{GetDimsubsetWithObsSparqlQuery}},
  \code{\link{GetDsdNameSparqlQuery}}
}

