% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCodeListSparqlQueryFull.R
\name{GetCodeListSparqlQueryFull}
\alias{GetCodeListSparqlQueryFull}
\title{SPARQL query for codelists in RDF data cube}
\usage{
GetCodeListSparqlQueryFull(forsparqlprefix, dsdName = NULL)
}
\arguments{
\item{forsparqlprefix}{string with prefix defintions}

\item{dsdName}{Dataset Descriptor Name, if NULL then return for all dsd in the cube}
}
\value{
SPARQL query
}
\description{
SPARQL query for codelists in RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetAttributesSparqlQuery}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetCodeListSpecSparqlQuery}},
  \code{\link{GetComponentSparqlQuery}},
  \code{\link{GetDimAttrSparqlQuery}},
  \code{\link{GetDimensionsSparqlQuery}},
  \code{\link{GetDimsubsetDescSparqlQuery}},
  \code{\link{GetDimsubsetWithObsSparqlQuery}},
  \code{\link{GetDsdNameSparqlQuery}}
}

