% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCodeListSpecSparqlQuery.R
\name{GetCodeListSpecSparqlQuery}
\alias{GetCodeListSpecSparqlQuery}
\title{SPARQL query for which codelists in RDF data cube}
\usage{
GetCodeListSpecSparqlQuery(forsparqlprefix, dsdName = NULL)
}
\arguments{
\item{forsparqlprefix}{string with prefix defintions}

\item{dsdName}{Dataset Descriptor Name, if NULL then return for all dsd in the cube}
}
\value{
SPARQL query
}
\description{
SPARQL query for which codelists in RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetAttributesSparqlQuery}},
  \code{\link{GetCodeListSparqlQueryFull}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetComponentSparqlQuery}},
  \code{\link{GetDimAttrSparqlQuery}},
  \code{\link{GetDimensionsSparqlQuery}},
  \code{\link{GetDimsubsetDescSparqlQuery}},
  \code{\link{GetDimsubsetWithObsSparqlQuery}},
  \code{\link{GetDsdNameSparqlQuery}}
}

