% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCommonPrefixDf.R
\name{GetCommonPrefixDf}
\alias{GetCommonPrefixDf}
\title{Get the common PREFIXES for a SPARQL query}
\usage{
GetCommonPrefixDf()
}
\value{
data.frame with columns prefix and namespace
}
\description{
Get the common PREFIXES for a SPARQL query
}

