% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetComponentSparqlQuery.R
\name{GetComponentSparqlQuery}
\alias{GetComponentSparqlQuery}
\title{SPARQL query for codelists in RDF data cube}
\usage{
GetComponentSparqlQuery(forsparqlprefix, dsdName)
}
\arguments{
\item{forsparqlprefix}{string with prefix defintions}

\item{dsdName}{Dataset Descriptor Name}
}
\value{
SPARQL query
}
\description{
SPARQL query for codelists in RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetAttributesSparqlQuery}},
  \code{\link{GetCodeListSparqlQueryFull}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetCodeListSpecSparqlQuery}},
  \code{\link{GetDimAttrSparqlQuery}},
  \code{\link{GetDimensionsSparqlQuery}},
  \code{\link{GetDimsubsetDescSparqlQuery}},
  \code{\link{GetDimsubsetWithObsSparqlQuery}},
  \code{\link{GetDsdNameSparqlQuery}}
}

