% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDimAttrSparqlQuery.R
\name{GetDimAttrSparqlQuery}
\alias{GetDimAttrSparqlQuery}
\title{SPARQL query for dimensions and attributes in RDF data cube}
\usage{
GetDimAttrSparqlQuery(forsparqlprefix)
}
\arguments{
\item{forsparqlprefix}{string with prefix defintions}
}
\value{
SPARQL query
}
\description{
SPARQL query for dimensions and attributes in RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetAttributesSparqlQuery}},
  \code{\link{GetCodeListSparqlQueryFull}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetCodeListSpecSparqlQuery}},
  \code{\link{GetComponentSparqlQuery}},
  \code{\link{GetDimensionsSparqlQuery}},
  \code{\link{GetDimsubsetDescSparqlQuery}},
  \code{\link{GetDimsubsetWithObsSparqlQuery}},
  \code{\link{GetDsdNameSparqlQuery}}
}

