% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDimensionsSparqlQuery.R
\name{GetDimensionsSparqlQuery}
\alias{GetDimensionsSparqlQuery}
\title{SPARQL query for dimensions in RDF data cube}
\usage{
GetDimensionsSparqlQuery(forsparqlprefix)
}
\arguments{
\item{forsparqlprefix}{PREFIX statements for SPARQL query}
}
\value{
SPARQL query
}
\description{
SPARQL query for dimensions in RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetAttributesSparqlQuery}},
  \code{\link{GetCodeListSparqlQueryFull}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetCodeListSpecSparqlQuery}},
  \code{\link{GetComponentSparqlQuery}},
  \code{\link{GetDimAttrSparqlQuery}},
  \code{\link{GetDimsubsetDescSparqlQuery}},
  \code{\link{GetDimsubsetWithObsSparqlQuery}},
  \code{\link{GetDsdNameSparqlQuery}}
}

