% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDimsubsetDescSparqlQuery.R
\name{GetDimsubsetDescSparqlQuery}
\alias{GetDimsubsetDescSparqlQuery}
\title{SPARQL query giving observations for a subset of observatition from RDF data cube}
\usage{
GetDimsubsetDescSparqlQuery(forsparqlprefix, domainName, dimensions)
}
\arguments{
\item{forsparqlprefix}{PREFIX part of SPARQL query}

\item{domainName}{domainName for the RRDF cube}

\item{dimensions}{dimensions}
}
\value{
SPARQL query
}
\description{
SPARQL query giving observations for a subset of observatition from RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetAttributesSparqlQuery}},
  \code{\link{GetCodeListSparqlQueryFull}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetCodeListSpecSparqlQuery}},
  \code{\link{GetComponentSparqlQuery}},
  \code{\link{GetDimAttrSparqlQuery}},
  \code{\link{GetDimensionsSparqlQuery}},
  \code{\link{GetDimsubsetWithObsSparqlQuery}},
  \code{\link{GetDsdNameSparqlQuery}}
}

