% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDimsubsetWithObsSparqlQuery.R
\name{GetDimsubsetWithObsSparqlQuery}
\alias{GetDimsubsetWithObsSparqlQuery}
\title{SPARQL query giving observations for a subset of observations from RDF data cube}
\usage{
GetDimsubsetWithObsSparqlQuery(forsparqlprefix, domainName, dimensionsorattrs)
}
\arguments{
\item{forsparqlprefix}{PREFIX part of SPARQL query}

\item{domainName}{domainName for the RRDF cube}

\item{dimensionsorattrs}{vector of dimensions or attributes}
}
\value{
SPARQL query
}
\description{
SPARQL query giving observations for a subset of observations from RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetAttributesSparqlQuery}},
  \code{\link{GetCodeListSparqlQueryFull}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetCodeListSpecSparqlQuery}},
  \code{\link{GetComponentSparqlQuery}},
  \code{\link{GetDimAttrSparqlQuery}},
  \code{\link{GetDimensionsSparqlQuery}},
  \code{\link{GetDimsubsetDescSparqlQuery}},
  \code{\link{GetDsdNameSparqlQuery}}
}

