% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDsdNameFromCube.R
\name{GetDsdNameFromCube}
\alias{GetDsdNameFromCube}
\title{Get dsdname for the RDF data cube
TODO: change the way domainname is used - should only use the dsdName}
\usage{
GetDsdNameFromCube(store, forsparqlprefixcommon = GetForSparqlPrefix())
}
\arguments{
\item{store}{the RDF store containing the datacube}

\item{forsparqlprefixcommon}{string with prefix defintions}
}
\value{
the dsdName for the cube
}
\description{
Get dsdname for the RDF data cube
TODO: change the way domainname is used - should only use the dsdName
}

