% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDsdNameSparqlQuery.R
\name{GetDsdNameSparqlQuery}
\alias{GetDsdNameSparqlQuery}
\title{SPARQL query for DataStructureDefinition in RDF data cube}
\usage{
GetDsdNameSparqlQuery(forsparqlprefixcommon = GetForSparqlPrefix())
}
\arguments{
\item{forsparqlprefixcommon}{string with prefix defintions}

\item{dsdName}{Dataset Descriptor Name}
}
\value{
SPARQL query
}
\description{
SPARQL query for DataStructureDefinition in RDF data cube
}
\seealso{
Other SPARQL.queries: \code{\link{GetAttributesSparqlQuery}},
  \code{\link{GetCodeListSparqlQueryFull}},
  \code{\link{GetCodeListSparqlQuery}},
  \code{\link{GetCodeListSpecSparqlQuery}},
  \code{\link{GetComponentSparqlQuery}},
  \code{\link{GetDimAttrSparqlQuery}},
  \code{\link{GetDimensionsSparqlQuery}},
  \code{\link{GetDimsubsetDescSparqlQuery}},
  \code{\link{GetDimsubsetWithObsSparqlQuery}}
}

