% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.recode.R
\name{ph.recode}
\alias{ph.recode}
\title{Return code for value using codelist}
\usage{
ph.recode(s, l)
}
\arguments{
\item{s}{A value interpreted as character}

\item{l}{A named list using the value as name as part of a URI}
}
\value{
s capitalized
}

