% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb.buildDSD.R
\name{qb.buildDSD}
\alias{qb.buildDSD}
\title{Add RDF Data Set Specification (DSD) to the rrdf store}
\usage{
qb.buildDSD(store, prefixlist, obsData, skeletonSource,
  dsdURIwoprefix = "dataset-demog", dsdName = "dsd-demog",
  extra = list(description =
  paste0("Cube with 6 Dimensions (factor, procedure, race, saffl, sex, trt01a),",
  "2 Attributes (denominator, unit) and 1 measure (measure)"), comment =
  paste0("Example Demographics data supplied by Ian Fleming via R. ",
  "All dimensions have a Codelist and Range specified. ",
  "Attributes applied from source data. ",
  "Attributes as VALUES instead of URIs."), label =
  "Demographics results data set.", distribution = "dataCubeFileName",
  obsfilename = "the name of the input file", obsDataSetName =
  "the name for the data set (D2RQ)",      title =
  "Demographics Analysis Results", PAVnodes = list(createdOn =
  gsub("(\\\\d\\\\d)$", ":\\\\1", strftime(Sys.time(),
  "\%Y-\%m-\%dT\%H:\%M:\%S\%z")), createdBy = "Tim Williams", pavVersion =
  "0.0.0", createdWith = paste0("R Version ", R.version$major, ".",
  R.version$minor, " Platform:", R.version$platform,
  " rrdfqbrnd0 package and dependencies"), providedBy =
  "PhUSE Results Metadata Working Group")), remote.endpoint = NULL,
  extension.rrdfqbcrnd0 = FALSE)
}
\arguments{
\item{store}{RRDF store}

\item{prefixlist}{prefix list}

\item{skeletonSource}{data.frame with skeleton source}

\item{dsdURIwoprefix}{DSD URI}

\item{dsdName}{DSD name}

\item{extra}{A list with member names: description, comment, label , distribution, obsfilename, title, PAVnodes, obsDataSetName}
}
\value{
Always TRUE
}
\description{
Add RDF Data Set Specification (DSD) to the rrdf store
}
\author{
Tim Williams, Marc Andersen
}

