% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb.buildObservations.R
\name{qb.buildObservations}
\alias{qb.buildObservations}
\title{Add observation for RDF Data Cube in rrdf store using values from data frame}
\usage{
qb.buildObservations(store, prefixlist, obsData, skeletonSource, dsdURIwoprefix,
  dsdName, recode.list, procedure2format)
}
\arguments{
\item{store}{RRDF store}

\item{prefixlist}{prefix list}

\item{skeletonSource}{data.frame with skeleton source}

\item{dsdURIwoprefix}{DSD URI}

\item{dsdName}{DSD name}

\item{recode.list}{A list of lists specifying how to recode the
value in the data frame. If NULL then the recode.list is generated
from the store}

\item{procedure2format}{A list specifying the format for the
descriptive statistics. If NULL then the default list is used}
}
\value{
Always TRUE
}
\description{
Add observation for RDF Data Cube in rrdf store using values from data frame
}

