% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb.buildSkeleton.R
\name{qb.buildSkeleton}
\alias{qb.buildSkeleton}
\title{Create RDF data cube skeleton from}
\usage{
qb.buildSkeleton(store, prefixlist, obsData, skeletonSource)
}
\arguments{
\item{store}{RRDF store}

\item{prefixlist}{prefix list}

\item{skeletonSource}{data.frame with skeleton source}
}
\value{
always TRUE
}
\description{
The skeletonSource is expected to contain compType corresponding to the compoents in
an RDF data cube: dimension, measure and attribute.
Consider: shorten code by introducing variables for properties etc. MJA 16-11-2014
}
\author{
Tim Williams, Marc Andersen
}

