% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrdfqb.R
\docType{data}
\name{qbIClist}
\alias{qbIClist}
\title{RDF data cube Integrity Constraints SPARQL queries}
\format{A list with SPARQL queries}
\description{
A dataset containing the RDF data cube Integrity Constraints
SPARQL queries as a list. The dataset is generated by the script
\file{data-raw/create-qb-IC-dataset.Rmd}
}
\references{
\url{http://www.w3.org/TR/2014/REC-vocab-data-cube-20140116/}
}
\keyword{datasets}

