% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunQbIC.R
\name{RunQbIC}
\alias{RunQbIC}
\title{Run the RDF data cube Integrity Constraints
The present version does not run IC-20 and IC-21.
Note TODO(mja): RRDF does not support the ASK query. This is handled
by converting ASK to SELECT.}
\usage{
RunQbIC(model, forsparqlprefix, doForIC = NULL)
}
\arguments{
\item{model}{a RRDF model containing one or more date cubes}

\item{forsparqlprefix}{string with the prefix to be used}

\item{doForIC}{character vector of specific integrity checks to run}
}
\value{
data.frame with columns ICtitle, ICfailobs
}
\description{
Run the RDF data cube Integrity Constraints
The present version does not run IC-20 and IC-21.
Note TODO(mja): RRDF does not support the ASK query. This is handled
by converting ASK to SELECT.
}

