% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/storeCodelist.from.df.R
\name{storeCodelist.from.df}
\alias{storeCodelist.from.df}
\title{Add data frame as codelist to RDF data cube in RDF store}
\usage{
storeCodelist.from.df(store, prefixlist, codeSource, codeType, nciDomainValue,
  dimName, underlDataSetName = NULL, remote.endpoint = NULL)
}
\arguments{
\item{store}{The name for the rrdf store}

\item{prefixlist}{R list of the prefixes}

\item{codeType}{Character "DATA" or "SDTM".
"DATA" to derive code list from the data.
"SDTM" to derive the code list from the rdf.cdisc documentation using a SPARQL query}

\item{nciDomainValue}{When codetype="SDTM" the nciDomain used for identifying the codelist}

\item{dimName}{the name of the dimension - for codeType="DATA" the name of the variable in the data frame ObsData}

\item{underlDataSetName}{underlying data set name. Used for finding name for D2RQ propertybridge. If NULL then not used.}

\item{remote.endpoint}{Used when codetype="SDTM" to give the URL for the remote endpoint. If NULL then the local rdf.cdisc.store from the environment is used.}

\item{obsData}{Data Frame with the data for which the code list is to be generated}
}
\value{
Alway TRUE - to be corrected
}
\description{
This function could be split into two function corresponding to the usage.
}
\author{
Tim Williams, Marc Andersen
}

